Liquibase Configuration added

+++ spring_security_demo/src/main/java/com/example/spring_security_demo/datasource/Student.java
@@ -14,7 +14,7 @@ public class Student extends BaseAuditingEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -23,16 +23,16 @@ public class Student extends BaseAuditingEntity {
     private String schoolName;
 
     @Column(name = "school_roll_no")
-    private Integer schoolRollNo;
+    private Long schoolRollNo;
 
     @Column(name = "class_id")
     private String classId;
 
     @Column(name = "roll_no")
-    private Integer rollNo;
+    private Long rollNo;
 
     @Column(name = "reg_no")
-    private Integer regNo;
+    private Long regNo;
 
     @Column(name = "verification_no")
     private String verificationNo;
@@ -40,7 +40,7 @@ public class Student extends BaseAuditingEntity {
     @Column(name = "marks")
     private Double marks;
 
-    public Student(String name, String schoolName, Integer schoolRollNo, String classId, Integer rollNo, Integer regNo, Double marks) {
+    public Student(String name, String schoolName, Long schoolRollNo, String classId, Long rollNo, Long regNo, Double marks) {
         this.name = name;
         this.schoolName = schoolName;
         this.schoolRollNo = schoolRollNo;
